Installed the ML Agents Toolkit Python Backend
cd <directory-of-project>
For Windows (you must have Python 3.6.x or higher installed):
python -m venv venv
venv\Scripts\activate
pip install -Iv mlagents==0.12.0
The package will overwrite your Physics collision matrix to ensure that Karts do not collide with each other.
This is used in training so we can have multiple karts on each track.
Remember to turn this back on if you wish for the karts to collide with each other.